Move `DataFrame` Protocol into `df.py`
